+++ src/main/java/chalmers/pimp/model/ModelImpl.java
@@ -276,7 +276,7 @@ final class ModelImpl implements IModel {
       var point = new Point(x, y);
 
       Point centerPoint = canvas.getActiveLayer().getCenterPoint();
-      double rotation = canvas.getActiveLayer().getRotation();
+      int rotation = canvas.getActiveLayer().getRotation();
       layerRotation.start(centerPoint, rotation, point, createSnapShot());
     }
   }
@@ -303,7 +303,7 @@ final class ModelImpl implements IModel {
   }
 
   @Override
-  public void rotateActiveLayer(double alpha) {
+  public void rotateActiveLayer(int alpha) {
     canvas.setActiveLayerRotation(alpha);
     notifyCanvasUpdateListeners();
   }
